requirements.txt
jsonargparse==4.28.0
lightning==2.2.3
seaborn==0.13.2
tabulate==0.9.0
termcolor==2.4.0
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
wandb==0.16.6
wget==3.2
ptflops==0.7.3
hydra-core==1.3.2